# backend/openapi.json
{
  "components": {
    "schemas": {
...
      "AuthorizeUrlOut": {
        "properties": {
          "url": {
            "title": "Url",
            "type": "string"
          }
        },
        "required": [
          "url"
        ],
        "title": "AuthorizeUrlOut",
        "type": "object"
      },
...
      "ExternalShareIn": {
        "description": "Send to either a saved target (target_id) or an ad-hoc one straight\nfrom the picker (target); exactly one must be provided.",
        "properties": {
          "article_id": {
            "title": "Article Id",
            "type": "integer"
          },
          "message": {
            "default": "",
            "maxLength": 4000,
            "title": "Message",
            "type": "string"
          },
          "target": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ShareTargetIn"
              },
              {
                "type": "null"
              }
            ]
          },
          "target_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Target Id"
          }
        },
        "required": [
          "article_id"
        ],
        "title": "ExternalShareIn",
        "type": "object"
      },
      "ExternalShareOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "platform": {
            "enum": [
              "slack",
              "teams"
            ],
            "title": "Platform",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "target_display": {
            "title": "Target Display",
            "type": "string"
          }
        },
        "required": [
          "id",
          "platform",
          "target_display",
          "status",
          "created_at"
        ],
        "title": "ExternalShareOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IntegrationStatusOut": {
        "properties": {
          "account_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Account Name"
          },
          "configured": {
            "title": "Configured",
            "type": "boolean"
          },
          "connected": {
            "title": "Connected",
            "type": "boolean"
          },
          "platform": {
            "enum": [
              "slack",
              "teams"
            ],
            "title": "Platform",
            "type": "string"
          },
          "status": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          },
          "workspace_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Workspace Name"
          }
        },
        "required": [
          "platform",
          "configured",
          "connected"
        ],
        "title": "IntegrationStatusOut",
        "type": "object"
      },
...
      "ShareTargetIn": {
        "properties": {
          "display_name": {
            "maxLength": 255,
            "minLength": 1,
            "title": "Display Name",
            "type": "string"
          },
          "external_id": {
            "maxLength": 255,
            "minLength": 1,
            "title": "External Id",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "default": {},
            "title": "Meta",
            "type": "object"
          },
          "platform": {
            "enum": [
              "slack",
              "teams"
            ],
            "title": "Platform",
            "type": "string"
          },
          "target_type": {
            "enum": [
              "channel",
              "group",
              "dm",
              "chat"
            ],
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "platform",
          "external_id",
          "display_name",
          "target_type"
        ],
        "title": "ShareTargetIn",
        "type": "object"
      },
      "ShareTargetOut": {
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "last_used_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Used At"
          },
          "meta": {
            "additionalProperties": true,
            "default": {},
            "title": "Meta",
            "type": "object"
          },
          "platform": {
            "enum": [
              "slack",
              "teams"
            ],
            "title": "Platform",
            "type": "string"
          },
          "target_type": {
            "enum": [
              "channel",
              "group",
              "dm",
              "chat"
            ],
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "id",
          "platform",
          "external_id",
          "display_name",
          "target_type"
        ],
        "title": "ShareTargetOut",
        "type": "object"
      },
...
      "TargetOptionOut": {
        "description": "One row in the live target picker (proxied from the platform API).",
        "properties": {
          "display_name": {
            "title": "Display Name",
            "type": "string"
          },
          "external_id": {
            "title": "External Id",
            "type": "string"
          },
          "meta": {
            "additionalProperties": true,
            "default": {},
            "title": "Meta",
            "type": "object"
          },
          "saved_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Saved Id"
          },
          "target_type": {
            "enum": [
              "channel",
              "group",
              "dm",
              "chat"
            ],
            "title": "Target Type",
            "type": "string"
          }
        },
        "required": [
          "external_id",
          "display_name",
          "target_type"
        ],
        "title": "TargetOptionOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "NewsRead API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/integrations": {
      "get": {
        "operationId": "integration_status_api_integrations_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationStatusOut"
                  },
                  "title": "Response Integration Status Api Integrations Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Integration Status",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/integrations/{platform}": {
      "delete": {
        "operationId": "disconnect_api_integrations__platform__delete",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "title": "Platform",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Disconnect",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/integrations/{platform}/authorize": {
      "get": {
        "operationId": "authorize_api_integrations__platform__authorize_get",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "title": "Platform",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthorizeUrlOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Authorize",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/integrations/{platform}/targets": {
      "get": {
        "description": "Live channel/chat list proxied from the platform, with saved ones marked.",
        "operationId": "search_targets_api_integrations__platform__targets_get",
        "parameters": [
          {
            "in": "path",
            "name": "platform",
            "required": true,
            "schema": {
              "title": "Platform",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "default": "",
              "title": "Q",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TargetOptionOut"
                  },
                  "title": "Response Search Targets Api Integrations  Platform  Targets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Search Targets",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/share-targets": {
      "get": {
        "operationId": "list_share_targets_api_share_targets_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ShareTargetOut"
                  },
                  "title": "Response List Share Targets Api Share Targets Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "List Share Targets",
        "tags": [
          "integrations"
        ]
      },
      "post": {
        "operationId": "save_share_target_api_share_targets_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ShareTargetIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ShareTargetOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Save Share Target",
        "tags": [
          "integrations"
        ]
      }
    },
    "/api/share-targets/{target_id}": {
      "delete": {
        "operationId": "delete_share_target_api_share_targets__target_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "target_id",
            "required": true,
            "schema": {
              "title": "Target Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Delete Share Target",
        "tags": [
          "integrations"
        ]
      }
    },
...
    "/api/shares/external": {
      "post": {
        "description": "Send one article to one target, synchronously \u2014 the user is waiting in\nthe composer and deserves a real success/failure, not a queue ticket.",
        "operationId": "send_external_share_api_shares_external_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExternalShareIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExternalShareOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Send External Share",
        "tags": [
          "integrations"
        ]
      }
    },
...
  }
}